+++ BidService/Application/Interfaces/IAuctionValidationService.cs
@@ -5,7 +5,7 @@ public interface IAuctionValidationService
     Task<AuctionValidationResult> ValidateAuctionForBidAsync(
         Guid auctionId,
         string bidder,
-        int bidAmount,
+        decimal bidAmount,
         CancellationToken cancellationToken = default);
 
     Task<AuctionInfo?> GetAuctionDetailsAsync(
@@ -24,8 +24,8 @@ public class AuctionValidationResult
     public bool IsValid { get; set; }
     public string? ErrorCode { get; set; }
     public string? ErrorMessage { get; set; }
-    public int CurrentHighBid { get; set; }
-    public int ReservePrice { get; set; }
+    public decimal CurrentHighBid { get; set; }
+    public decimal ReservePrice { get; set; }
     public DateTimeOffset? AuctionEnd { get; set; }
     public string? Seller { get; set; }
     public string? Status { get; set; }
@@ -37,9 +37,9 @@ public class AuctionInfo
     public string Title { get; set; } = string.Empty;
     public string Seller { get; set; } = string.Empty;
     public string? Winner { get; set; }
-    public int CurrentHighBid { get; set; }
-    public int ReservePrice { get; set; }
-    public int? BuyNowPrice { get; set; }
+    public decimal CurrentHighBid { get; set; }
+    public decimal ReservePrice { get; set; }
+    public decimal? BuyNowPrice { get; set; }
     public DateTimeOffset AuctionEnd { get; set; }
     public string Status { get; set; } = string.Empty;
     public bool IsBuyNowAvailable { get; set; }
